Refactor to make pylint happy
